feat: Convert library to Deno module
This commit converts the existing Node.js module to a Deno module. The
JavaScript file `lib/index.js` is deleted and replaced with a TypeScript
file `mod.ts`. The corresponding test file is also created as `mod_test.ts`.
The Deno specific files `deno.json` and `deno.lock` are added. The
unnecessary TypeScript definition files and tests are removed.